[Cli] Moving InputHandler classes to a new Server\Handler namespace for better organization.
