+++ tests/Unit/Application/ApplicationTest.php
@@ -40,7 +40,7 @@ use Valkyrja\Cli\Routing\Collection\Contract\Collection as CliRoutingCollection;
 use Valkyrja\Cli\Routing\Collector\Contract\Collector as CliRoutingCollector;
 use Valkyrja\Cli\Routing\Data as CliData;
 use Valkyrja\Cli\Routing\Dispatcher\Contract\Router as CliRoutingRouter;
-use Valkyrja\Cli\Server\Contract\InputHandler as InputHandlerContract;
+use Valkyrja\Cli\Server\Handler\Contract\InputHandler as InputHandlerContract;
 use Valkyrja\Cli\Server\Middleware\LogThrowableCaughtMiddleware as CliLogThrowableCaughtMiddleware;
 use Valkyrja\Container\Collector\Contract\Collector as ContainerCollector;
 use Valkyrja\Container\Data as ContainerData;
