+++ tests/Unit/Cli/Server/Provider/ServiceProviderTest.php
@@ -19,8 +19,8 @@ use Valkyrja\Cli\Middleware\Handler\Contract\ExitedHandler;
 use Valkyrja\Cli\Middleware\Handler\Contract\InputReceivedHandler;
 use Valkyrja\Cli\Middleware\Handler\Contract\ThrowableCaughtHandler;
 use Valkyrja\Cli\Routing\Dispatcher\Contract\Router;
-use Valkyrja\Cli\Server\Contract\InputHandler as InputHandlerContract;
-use Valkyrja\Cli\Server\InputHandler;
+use Valkyrja\Cli\Server\Handler\Contract\InputHandler as InputHandlerContract;
+use Valkyrja\Cli\Server\Handler\InputHandler;
 use Valkyrja\Cli\Server\Middleware\LogThrowableCaughtMiddleware;
 use Valkyrja\Cli\Server\Provider\ServiceProvider;
 use Valkyrja\Log\Logger\Contract\Logger;
